tidy: better tx signature
By adding the lifetime on `run` clients can more easily capture things in the scope of the closure.
It's still kind of tricky because of `FnMut` and the fact that we can call the closure multiple times, but this makes it pretty ergonomic.

Also, this makes `Tx` visible. This might not be the best for testing, but it reduces the indirection and simplifies the signature.